Revert "feat: maintain scores"
